fix authentication and ldap messages

+++ backend/managementUsers/views.py
@@ -5,7 +5,7 @@ from django.utils.translation import gettext_lazy as _
 from django.core.exceptions import ValidationError
 from backend.managementUsers.models import Profile, User
 from backend.managementUsers.serializers import PermissionSerializer, ProfileSerializer, UserSerializerGet, UserSerializerPost, UserSerializerPostWithoutGroupAndPermission
-from backend.managementUsers.functions import add_user_group, add_mail_spool, add_user, reset_password_by_admin_in_system, change_username, check_same_groupname_with_username, delete_user_group, delete_user_in_system, get_uid_user, reset_password, reset_password_by_admin_in_system, username_exists, valid_input, valid_password
+from backend.managementUsers.functions import add_user_group, add_mail_spool, add_user, reset_password_by_admin_in_system, change_username, check_same_groupname_with_username, delete_user_group, delete_user_in_system, get_uid_user, reset_password, reset_password_by_admin_in_system, username_exists, valid_input, valid_password,delete_directory
 from backend.managementGroup.serializers import GroupSerializer
 from backend.managementGroup.functions import change_groupname_username, getGroupNameById, getUidGroup
 from backend.managementGroup.models import Group
@@ -38,6 +38,7 @@ CONSTANT_METHOD_ADD_USER_EMAIL_SYSTEM = _("with simple System email")
 CONSTANT_OR = _("or")
 CONSTANT_AND = _("and")
 CONSTANT_LANGUAGE = _('Language')
+CONSTANT_LDAP_UNREACHABLE=_("Directory Server unreachable")
 # Success messages
 SUCCESS_MESSAGES_CREATING = _("is created")
 SUCCESS_MESSAGES_DELETING = _("is deleted")
@@ -49,7 +50,6 @@ ERROR_MESSAGES_UPDATING = _("Error in updating")
 ERROR_MESSAGES_RESET = _("Error in reset")
 ERROR_MESSAGES_EXISTANT = _("already exist")
 ERROR_MESSAGES_INEXISTANT = _("does not exist")
-ERROR_MESSAGES_INVALID_CREDENTIALS = _("Invalid credentials")
 ERROR_MESSAGES_INVALID_PASSWORD = _("Invalid password")
 ERROR_MESSAGES_CONNECTION = _("Error connecting to directory server")
 
@@ -156,10 +156,9 @@ def create_user(request):
                         return JsonResponse({'msg': ERROR_MESSAGES_INVALID_PASSWORD}, status=400)   
                     
                 except ldap.SERVER_DOWN:
-                # LDAP authentication failed
-                    return JsonResponse({'msg': ERROR_MESSAGES_INVALID_CREDENTIALS}, status=400)
+                    return JsonResponse({'msg': f"{CONSTANT_LDAP_UNREACHABLE}"}, status=400)
                 except ldap.LDAPError:
-                    return JsonResponse({'msg': ERROR_MESSAGES_CONNECTION}, status=400)
+                    return JsonResponse({'msg': f"{ERROR_MESSAGES_CONNECTION}"}, status=400)
             else:
                 return JsonResponse({'msg': f"{CONSTANT_DIRECTORY_SERVER} {ERROR_MESSAGES_INEXISTANT}"}, status=400)
        
@@ -247,11 +246,13 @@ def delete_user(request, id):
     group = Group.objects.filter(groupname=user.username)
     # # Execute the command on the remote machine
     _, stderr = delete_user_in_system(user.username)
+    _, stderr_dir = delete_directory(user.username)
     # # convert the stderr stream to a string
     if stderr == "":
-        user.delete()
-        group.delete()
-        return JsonResponse({"msg": f"{user.username} {SUCCESS_MESSAGES_DELETING}"})
+        if stderr_dir == "":
+            user.delete()
+            group.delete()
+            return JsonResponse({"msg": f"{user.username} {SUCCESS_MESSAGES_DELETING}"})
     return JsonResponse({"error": f"{ERROR_MESSAGES_DELETING} {CONSTANT_USER}"}, status=400)
 
 
@@ -315,9 +316,9 @@ def modify_user(request, id):
                     return JsonResponse({'msg': ERROR_MESSAGES_INVALID_PASSWORD}, status=400)    
             except ldap.SERVER_DOWN:
                 # LDAP authentication failed
-                return JsonResponse({'msg': ERROR_MESSAGES_INVALID_CREDENTIALS}, status=400)        
+                return JsonResponse({'msg':f"{CONSTANT_LDAP_UNREACHABLE}"}, status=400)        
             except ldap.LDAPError:
-                return JsonResponse({'msg': ERROR_MESSAGES_CONNECTION}, status=400)  
+                return JsonResponse({'msg':f"{ERROR_MESSAGES_CONNECTION}"}, status=400)  
             
         if User.objects.filter(email=data['email']).exclude(id=id).exists():
             return JsonResponse({"msg": f"Email {ERROR_MESSAGES_EXISTANT}"}, status=400)
